Métodos serializers refatorados.
