+++ utils/serializers.py
@@ -1,0 +1,22 @@
+import json
+
+def serialize_get_customer_list(query_result):
+    serialized_data = [
+        {
+            'id': customer[0],
+            'name': customer[1],
+            'email': customer[2]
+        }
+        for customer in query_result
+    ]
+
+    return json.dumps(serialized_data)
+
+def serialize_get_customer(query_result):
+    serialized_data = {
+        'id': query_result[0],
+        'name': query_result[1],
+        'email': query_result[2]
+    }
+
+    return json.dumps(serialized_data)
